py/requirements.txt
face-recognition==1.3.0
face-recognition-models==0.3.0
nav-msgs==4.2.3
numpy==1.21.5
rcl-interfaces==1.2.1
rclpy==3.3.10
rcutils==5.1.3
requests==2.25.1
rmw-dds-common==1.6.0
ros2action==0.18.7
ros2bag==0.15.8
ros2cli==0.18.7
ros2component==0.18.7
ros2doctor==0.18.7
ros2interface==0.18.7
ros2launch==0.19.6
ros2lifecycle==0.18.7
ros2multicast==0.18.7
ros2node==0.18.7
ros2param==0.18.7
ros2pkg==0.18.7
ros2run==0.18.7
ros2service==0.18.7
ros2topic==0.18.7
rosbag2-interfaces==0.15.8
rosbag2-py==0.15.8
rosdistro-modules==0.9.0
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.1
rosidl-typesupport-fastrtps-cpp==2.2.1
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rospkg-modules==1.5.0
rpyutils==0.2.1
tf-transformations==1.0.0
tf2-geometry-msgs==0.25.4
tf2-kdl==0.25.4
tf2-msgs==0.25.4
tf2-py==0.25.4
tf2-ros-py==0.25.4
tf2-tools==0.25.4
Twisted==22.1.0

